+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- CLI and library functions to run a cirrus workflow, and collect its output.
+  Also adds a `call` command which is similar to `exec` but uses a
+  subprocess. And adds `invoke_lambda` which invokes lambdas that are part of
+  the cirrus deployment.  ([#4](https://github.com/cirrus-geo/cirrus-mgmt/pull/4)
+
+
 ## [v0.1.0] -
 
 Initial release